requirements.txt
beautifulsoup4==4.14.2
ncatbot==4.2.6
numpy==2.3.4
opencv_python==4.12.0.88
pandas==2.3.3
Pillow==12.0.0
pydantic==2.12.2
pyparsing==3.2.5
Requests==2.32.5
